refactor: add doltlite_branch_int.h for branch/checkout split
Scaffold a private header and promote BranchMutationCtx,
branchNameEmpty, and mutateBranchRef so an upcoming checkout
translation unit can call them. Wire into EXTRA and amalgamation
available_hdr. No checkout body moves yet.
