+++ src/doltlite_branch_int.h
@@ -1,0 +1,22 @@
+#ifndef DOLTLITE_BRANCH_INT_H
+#define DOLTLITE_BRANCH_INT_H
+
+/* Private declarations shared by the branch/checkout implementation modules. */
+
+#include "sqliteInt.h"
+#include "prolly_hash.h"
+#include "chunk_store.h"
+#include "doltlite_internal.h"
+
+typedef struct BranchMutationCtx BranchMutationCtx;
+struct BranchMutationCtx {
+  const char *zName;
+  ProllyHash head;
+  int isDelete;
+  int force;
+};
+
+int branchNameEmpty(const char *zName);
+int mutateBranchRef(sqlite3 *db, ChunkStore *cs, void *pArg);
+
+#endif /* DOLTLITE_BRANCH_INT_H */
